Lab-04_VLAN_ACL_DNS
This lab simulates a small business network with two VLANs (Staff and Guest), using DHCP and DNS services. An Access Control List (ACL) is implemented to restrict HTTP access from the Guest VLAN to the Staff VLAN and the internet. Only the Staff VLAN can access the simulated internet (via a DNS server).